==> analysis/5_spatial_vis.Rmd <==
---
title: "IMC data analysis workflow"
author: "Tess and Tural"
date: "University of Bern, DBMR, Visceral and Transplantation Surgery lab"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Spatial analysis 

```{r setup, include=FALSE}
#Setup knitr, incuding path
knitr::opts_chunk$set(echo = TRUE)

#Important - this path is relative to your dataset, set working directory
knitr::opts_knit$set(root.dir = "C:/Users/IMC Platform laptop/Users2/Tess/SteinbockCellSegmentation/221130_haibin/230515_workflowr_haibin")

#Set cache
knitr::opts_chunk$set(autodep = TRUE, cache = TRUE, cache.lazy = FALSE)

#To make figures
knitr::opts_chunk$set(dev = "png")
```

```{css, echo=FALSE}
.scroll-100 {
  max-height: 100px;
  overflow-y: auto;
  background-color: inherit;
}
```

### Load packages 

```{r load-libs, message=FALSE}
library(SpatialExperiment)
library(imcRtools)
library(ggplot2)
library(viridis)
library(igraph)
library(pheatmap)
library(tidyverse)
library(lisaClust)
library(spicyR)
library(circlize)
library(RColorBrewer)
library(cytomapper)
library(patchwork)
library(tidyverse)
library(scales)
library(dittoSeq)
library(BiocParallel)
```

### Load data

```{r load data, eval=FALSE, include=FALSE}
spe <- readRDS(file.path("output", "spe3.rds"))

```

### Spatial interaction graphs (only run once- very long)

```{r build_spatial_graph, eval=FALSE, include=FALSE}
set.seed(322)
spe <- buildSpatialGraph(spe, img_id = "sample_id", type = "knn", k = 20)
spe <- buildSpatialGraph(spe, img_id = "sample_id", type = "expansion", threshold = 20)
spe <- buildSpatialGraph(spe, img_id = "sample_id", type = "delaunay", max_dist = 50)

colPairNames(spe)
```

### Save data with interaction graphs

```{r save spe4, eval=FALSE, include=FALSE}
saveRDS(spe, file.path("output", "spe4.rds"))
```

### Load data

```{r load data2}
spe <- readRDS(file.path("output", "spe4.rds"))

```

### Define colors

```{r set colors}

color_vectors <- list()

#For the ROI
ROI <- setNames(dittoColors(reps = 1)[seq_along(unique(spe$sample_id))], 
                unique(spe$ROI))
#For the patient ID
patient_id <- setNames(brewer.pal(length(unique(spe$patient_id)), name = "Set1"), 
                unique(spe$patient_id))
#For the sample ID
sample_id <- setNames(dittoColors(reps = 1)[seq_along(unique(spe$sample_id))], 
                unique(spe$sample_id))
#For the condition
condition <- setNames(brewer.pal(length(unique(spe$condition)), name = "Set2"), 
                unique(spe$condition))
#For the annotation
annotation <- setNames(brewer.pal(length(unique(spe$annotation)), name = "Set3"), 
                unique(spe$annotation))
#For cell type
celltype <- setNames(dittoColors(reps = 1)[seq_along(unique(spe$sample_id))], 
                unique(spe$celltype))

color_vectors$ROI <- ROI
color_vectors$patient_id <- patient_id
color_vectors$sample_id <- sample_id
color_vectors$condition <- condition
color_vectors$annotation <- annotation
color_vectors$celltype <- celltype

metadata(spe)$color_vectors <- color_vectors
```

### Interaction graphs 

```{r steinbock_interaction}

# steinbock interaction graph 
plotSpatial(spe[,spe$sample_id == "221031_day1_Haibin_1_6"], 
            node_color_by = "celltype", 
            img_id = "sample_id", 
            draw_edges = TRUE, 
            colPairName = "neighborhood", 
            nodes_first = FALSE, 
            edge_color_fix = "grey") + 
    scale_color_manual(values = metadata(spe)$color_vectors$celltype) +
    ggtitle("steinbock interaction graph")
```

```{r KNN_interaction}

# knn interaction graph 
plotSpatial(spe[,spe$sample_id == "221031_day1_Haibin_1_6"], 
            node_color_by = "celltype", 
            img_id = "sample_id", 
            draw_edges = TRUE, 
            colPairName = "knn_interaction_graph", 
            nodes_first = FALSE,
            edge_color_fix = "grey") + 
    scale_color_manual(values = metadata(spe)$color_vectors$celltype) +
    ggtitle("knn interaction graph")
```

```{r expansion_interaction}

# expansion interaction graph 
plotSpatial(spe[,spe$sample_id == "221031_day1_Haibin_1_6"], 
            node_color_by = "celltype", 
            img_id = "sample_id", 
            draw_edges = TRUE, 
            colPairName = "expansion_interaction_graph", 
            nodes_first = FALSE, 
            directed = FALSE,
            edge_color_fix = "grey") + 
    scale_color_manual(values = metadata(spe)$color_vectors$celltype) +
    ggtitle("expansion interaction graph")
```

```{r delaunay_interaction}
# delaunay interaction graph 
plotSpatial(spe[,spe$sample_id == "221031_day1_Haibin_1_6"], 
            node_color_by = "celltype", 
            img_id = "sample_id", 
            draw_edges = TRUE, 
            colPairName = "delaunay_interaction_graph", 
            nodes_first = FALSE,
            edge_color_fix = "grey") + 
    scale_color_manual(values = metadata(spe)$color_vectors$celltype) +
    ggtitle("delaunay interaction graph")
```

```{r interaction_targetexprs}
plotSpatial(spe[,spe$sample_id == "221031_day1_Haibin_1_6"], 
            node_color_by = "E-cadherin", 
            assay_type = "exprs",
            img_id = "sample_id", 
            draw_edges = TRUE, 
            colPairName = "expansion_interaction_graph", 
            nodes_first = FALSE, 
            node_size_by = "area", 
            directed = FALSE,
            edge_color_fix = "grey") + 
    scale_size_continuous(range = c(0.1, 2)) +
    ggtitle("E-cadherin expression")
```

### Plot cell types

```{r plot_cell_type_all, fig.width= 20, fig.height= 25}
plotSpatial(spe, 
            node_color_by = "celltype", 
            img_id = "sample_id", 
            node_size_fix = 0.5) 
    scale_color_manual(values = metadata(spe)$color_vectors$celltype)
```

### Spatial community analysis

```{r new community steinbock interactions}
spe$tumor_stroma <- ifelse(spe$celltype == "Tumor", "Tumor", "Stroma")

spe <- detectCommunity(spe, 
                       colPairName = "neighborhood", 
                       size_threshold = 10,
                       group_by = "tumor_stroma",
                       BPPARAM = SerialParam(RNGseed = 220819))
```

### Plot communities of tumor and notumor

```{r plot tumor communities, fig.width= 20, fig.height= 25}
plotSpatial(spe[,spe$celltype == "Tumor"], 
            node_color_by = "spatial_community", 
            img_id = "sample_id", 
            node_size_fix = 0.5) +
    theme(legend.position = "none") +
    ggtitle("Spatial tumor communities") 
    # scale_color_manual(values = rev(colors()))
```

### Plot nontumor communities

```{r plot nontumor communities, fig.width= 20, fig.height= 25}
plotSpatial(spe[,spe$celltype != "Tumor"], 
            node_color_by = "spatial_community", 
            img_id = "sample_id", 
            node_size_fix = 0.5) +
    theme(legend.position = "none") +
    ggtitle("Spatial non-tumor communities") 
    # scale_color_manual(values = rev(colors()))
```

### Plot nontumor community heatmap

```{r community_nontumor_heatmap}
for_plot <- prop.table(table(spe[,spe$celltype != "Tumor"]$spatial_community, 
                             spe[,spe$celltype != "Tumor"]$celltype), margin = 1)

pheatmap(for_plot, color = viridis(100), show_rownames = FALSE)
```

```{r community_all_heatmap}
for_plot <- prop.table(table(spe$spatial_community, 
                             spe$celltype), margin = 1)

pheatmap(for_plot, color = viridis(100), show_rownames = FALSE)
```

### Cellular neighborhood analysis
### Cn_1 is neighborhoods made by cell types

```{r knn_interaction_graph_prep}
# By celltypes
spe <- aggregateNeighbors(spe, colPairName = "knn_interaction_graph", 
                          aggregate_by = "metadata", count_by = "celltype")

set.seed(322)
cn_cell1 <- kmeans(spe$aggregatedNeighbors, centers = 4)
set.seed(322)
cn_cell2 <- kmeans(spe$aggregatedNeighbors, centers = 5)
set.seed(322)
cn_cell3 <- kmeans(spe$aggregatedNeighbors, centers = 6)
```

```{r knn_interaction_graph, fig.width= 20, fig.height= 25}


spe$cn_celltypes <- as.factor(cn_cell3$cluster)

plotSpatial(spe, 
            node_color_by = "cn_celltypes", 
            img_id = "sample_id", 
            node_size_fix = 0.5) +
    scale_color_brewer(palette = "Set3")
```

```{r knn_interaction_heatmap}
for_plot <- colData(spe) %>% as_tibble() %>%
    group_by(cn_celltypes, celltype) %>%
    summarize(count = n()) %>%
    mutate(freq = count / sum(count)) %>%
    pivot_wider(id_cols = cn_celltypes, names_from = celltype, 
                values_from = freq, values_fill = 0) %>%
    ungroup() %>%
    select(-cn_celltypes)

pheatmap(for_plot, color = colorRampPalette(c("dark blue", "white", "dark red"))(100), 
         scale = "column")
```

#Double split barplots

```{r celltype communities by condition and annotation, fig.width= 10, fig.height= 12}
plotdata <- as.data.frame(colData(spe)[, c("cn_celltypes", "annotation", "condition")])
plotdata <- as.data.frame(table(plotdata)) 

ggplot(plotdata, aes(x = annotation, y = Freq, fill = condition)) + 
  geom_col() +
  facet_grid( ~ cn_celltypes)

ggplot(plotdata, aes(x = annotation, y = Freq, fill = condition)) + 
  geom_col() +
  facet_wrap( ~ cn_celltypes, scales = "free", ncol = 14)

```


### Cn2 is neighborhoods made by target expression in 20 nearest neighbors

```{r neighborhoods_by_exprs, fig.width= 20, fig.height= 25}
# By expression
spe <- aggregateNeighbors(spe, colPairName = "knn_interaction_graph", 
                          aggregate_by = "expression", assay_type = "exprs",
                          subset_row = rowData(spe)$use_channel)
cn_exp <- kmeans(spe$mean_aggregatedExpression, centers = 6)
spe$cn_expression <- as.factor(cn_exp$cluster)

plotSpatial(spe, 
            node_color_by = "cn_expression", 
            img_id = "sample_id", 
            node_size_fix = 0.5) +
    scale_color_brewer(palette = "Set3")
```

```{r neighborhoods_by_expression_heatmap}
for_plot <- colData(spe) %>% as_tibble() %>%
    group_by(cn_expression, celltype) %>%
    summarize(count = n()) %>%
    mutate(freq = count / sum(count)) %>%
    pivot_wider(id_cols = cn_expression, names_from = celltype, 
                values_from = freq, values_fill = 0) %>%
    ungroup() %>%
    select(-cn_expression)

pheatmap(for_plot, color = colorRampPalette(c("dark blue", "white", "dark red"))(100), 
         scale = "column")
```

### Cellular neighborhood analysis: lisaClust package computes local indicators of spatial ###associations (LISA) functions and cluster cells based on those

```{r lisaClust_setup, eval=FALSE, include=FALSE}
cells <- data.frame(row.names = colnames(spe))
cells$ObjectNumber <- spe$ObjectNumber
cells$ImageNumber <- spe$sample_id
cells$AreaShape_Center_X <- spatialCoords(spe)[,"Pos_X"]
cells$AreaShape_Center_Y <- spatialCoords(spe)[,"Pos_Y"]
cells$cellType <- spe$celltype

lisa_sc <- SegmentedCells(cells, cellProfiler = TRUE)

lisa_sc
```

### Does not work yet

```{r lisaClust_plot, eval=FALSE, include=FALSE}

lisaCurves <- lisa(lisa_sc, Rs = c(10, 20, 50))

# Set NA to 0
lisaCurves[is.na(lisaCurves)] <- 0

lisa_clusters <- kmeans(lisaCurves, centers = 6)$cluster

spe$lisa_clusters <- as.factor(lisa_clusters)

plotSpatial(spe, 
            node_color_by = "lisa_clusters", 
            img_id = "sample_id", 
            node_size_fix = 0.5) +
    scale_color_brewer(palette = "Set3")
```

```{r lisaClust_heatmap, eval=FALSE, include=FALSE}
for_plot <- colData(spe) %>% as_tibble() %>%
    group_by(lisa_clusters, celltype) %>%
    summarize(count = n()) %>%
    mutate(freq = count / sum(count)) %>%
    pivot_wider(id_cols = lisa_clusters, names_from = celltype, 
                values_from = freq, values_fill = 0) %>%
    ungroup() %>%
    select(-lisa_clusters)

pheatmap(for_plot, color = colorRampPalette(c("dark blue", "white", "dark red"))(100), 
         scale = "column")
```

### Spatial context analysis (takes a while)

```{r spatial_context_analysis, fig.width= 25, fig.height= 25}
# Generate k-nearest neighbor graph for SC detection (k=40) 
spe <- buildSpatialGraph(spe, img_id = "sample_id", 
                         type = "knn", 
                         name = "knn_spatialcontext_graph", 
                         k = 40)

# Aggregate based on clustered_neighbors
spe <- aggregateNeighbors(spe, 
                          colPairName = "knn_spatialcontext_graph",
                          aggregate_by = "metadata",
                          count_by = "cn_celltypes",
                          name = "aggregatedNeighborhood")

# Detect spatial contexts
spe <- detectSpatialContext(spe, 
                            entry = "aggregatedNeighborhood",
                            threshold = 0.90,
                            name = "spatial_context")

# Define SC color scheme
col_SC <- setNames(colorRampPalette(brewer.pal(9, "Paired"))(length(unique(spe$spatial_context))), 
                   sort(unique(spe$spatial_context)))

# Visualize spatial contexts on images
plotSpatial(spe, 
            node_color_by = "spatial_context", 
            img_id = "sample_id", 
            node_size_fix = 0.5, 
            colPairName = "knn_spatialcontext_graph") +
    scale_color_manual(values = col_SC)
```

```{r compare_CN_SC, fig.width= 20, fig.height= 10}

# Compare CN and SC for one patient 
p1 <- plotSpatial(spe[,spe$sample_id == "221031_day1_Haibin_1_6"], 
            node_color_by = "cn_celltypes", 
            img_id = "sample_id", 
            node_size_fix = 0.5, 
            colPairName = "knn_interaction_graph") +
    scale_color_brewer(palette = "Set3")

p2 <- plotSpatial(spe[,spe$sample_id == "221031_day1_Haibin_1_6"], 
            node_color_by = "spatial_context", 
            img_id = "sample_id", 
            node_size_fix = 0.5, 
            colPairName = "knn_spatialcontext_graph") +
    scale_color_manual(values = col_SC, limits = force)

p1 + p2
```

```{r filter_SC, fig.width= 20, fig.height= 10}
## Filter spatial contexts
# By number of group entries
spe <- filterSpatialContext(spe, 
                            entry = "spatial_context",
                            group_by = "patient_id", 
                            group_threshold = 3)

plotSpatial(spe, 
            node_color_by = "spatial_context_filtered", 
            img_id = "sample_id", 
            node_size_fix = 0.5, 
            colPairName = "knn_spatialcontext_graph") +
    scale_color_manual(values = col_SC, limits = force)
```

```{r filter_SC_bycellnumber, fig.width= 20, fig.height= 10}
# By number of group entries and total number of cells
spe <- filterSpatialContext(spe, 
                            entry = "spatial_context",
                            group_by = "patient_id", 
                            group_threshold = 3,
                            cells_threshold = 100)

plotSpatial(spe, 
            node_color_by = "spatial_context_filtered", 
            img_id = "sample_id", 
            node_size_fix = 0.5, 
            colPairName = "knn_spatialcontext_graph") +
    scale_color_manual(values = col_SC, limits = force)
```

```{r sc_graphs}
## Plot spatial context graph 

# Colored by name and size by n_cells
plotSpatialContext(spe, 
                   entry = "spatial_context_filtered",
                   group_by = "sample_id",
                   node_color_by = "name",
                   node_size_by = "n_cells",
                   node_label_color_by = "name")
```

```{r SC_graphs_ncellsngroup_color}
# Colored by n_cells and size by n_group                   
plotSpatialContext(spe, 
                   entry = "spatial_context_filtered",
                   group_by = "sample_id",
                   node_color_by = "n_cells",
                   node_size_by = "n_group",
                   node_label_color_by = "n_cells") +
  scale_color_viridis()
```

### Patch detection 

```{r patch detection, fig.width= 20, fig.height= 20}
spe <- patchDetection(spe, 
                      patch_cells = spe$celltype == "Tumor",
                      img_id = "sample_id",
                      expand_by = 1,
                      min_patch_size = 10,
                      colPairName = "neighborhood")

plotSpatial(spe, 
            node_color_by = "patch_id", 
            img_id = "sample_id", 
            node_size_fix = 0.5) +
    theme(legend.position = "none") 
    # scale_color_manual(values = colors())
```

```{r T cell infiltration, fig.width= 20, fig.height= 10}
colData(spe) %>% as_tibble() %>%
    group_by(patch_id, sample_id) %>%
    summarize(Tcell_count = sum(celltype == "CD8" | celltype == "CD4"),
              patch_size = n(),
              Tcell_freq = Tcell_count / patch_size) %>%
    ggplot() +
        geom_point(aes(log10(patch_size), Tcell_freq, color = sample_id)) +
    theme_classic()
```

```{r patch_size}
patch_size <- patchSize(spe, "patch_id")

patch_size <- merge(patch_size, 
                    colData(spe)[match(patch_size$patch_id, spe$patch_id),], 
                    by = "patch_id")

ggplot(as.data.frame(patch_size)) + 
    geom_boxplot(aes(patient_id, log10(size))) +
    geom_point(aes(patient_id, log10(size)))
```

```{r min_dist_cells, fig.width= 20, fig.height= 20}
spe <- minDistToCells(spe, 
                      x_cells = !is.na(spe$patch_id), 
                      img_id = "sample_id")

plotSpatial(spe, 
            node_color_by = "distToCells", 
            img_id = "sample_id", 
            node_size_fix = 0.5) +
    scale_color_gradient2(low = "dark blue", mid = "white", high = "dark red")
```

### Distance histograms 

```{r dist_histograms}
library(ggridges)
ggplot(as.data.frame(colData(spe))) + 
    geom_density_ridges(aes(distToCells, celltype, fill = celltype)) +
    geom_vline(xintercept = 0, color = "dark red", linewidth = 2) +
    scale_fill_manual(values = metadata(spe)$color_vectors$celltype)
```

### Interaction analysis 

```{r interactions_setup}
out <- testInteractions(spe, 
                        group_by = "sample_id",
                        label = "celltype", 
                        colPairName = "neighborhood")

head(out)
```

```{r interactions_heatmap}

out %>% as_tibble() %>%
    group_by(from_label, to_label) %>%
    summarize(sum_sigval = sum(sigval, na.rm = TRUE)) %>%
    ggplot() +
        geom_tile(aes(from_label, to_label, fill = sum_sigval)) +
        scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
        theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r interactions_heatmap_patch}
out2 <- testInteractions(spe, 
                        group_by = "sample_id",
                        label = "celltype", 
                        colPairName = "neighborhood",
                        method = "patch", 
                        patch_size = 3)

out2 %>% as_tibble() %>%
    group_by(from_label, to_label) %>%
    summarize(sum_sigval = sum(sigval, na.rm = TRUE)) %>%
    ggplot() +
        geom_tile(aes(from_label, to_label, fill = sum_sigval)) +
        scale_fill_gradient2(low = muted("blue"), mid = "white", high = muted("red")) +
        theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

### Save data 

```{r save spe5}
saveRDS(spe, file.path("output", "spe5.rds"))
```
